Writer for datapoints and spans
These are components that accept datapoints/spans from a channel and
deal with buffering/sending them in a manner that provides a good
trade-off between latency and number of requests.

We have implemented this same thing many times so it seems good to put
it in a common place where it is easily reused.

The first implementation uses a ring buffer to minimize allocations.
This has been in use in the Smart Agent for many months now without
issue.  It was largely copied from there.