Update spans validator to accept empty body as a valid response
Updated span validator function to accept empty body in response so it
can export spans in zipkin format to any valid zipkin destination in
addition to the SignalFx backend.

Alternatively, we can make the spans validator pluggable in HTTP Sink so
a separate specific Zipkin exporter can override the behavior.