Extend SFXAPIError to include endpoint field
This change extends the existing SFXAPIError struct by adding an
endpoint field This allows for custom error handlers that can behave
differently depending on the endpoint associated with the
error. In the AsyncMultiTokenSink case, this error handler is only
called when the underlying client can't successfully call the ingestion
endpoint after the configured number of retries.

An example of where this could be useful:

The token for SignalFx ingestion is being throttled, so report the
error directly to cloudwatch so we can alert on it, we may want to
set up an alert differently for a datapoint throttling issue versus an
event throttling issue.

update go.mod

remove token from SFXAPIError

remove adampetrovic mod path