web: adding service state support
currently web CloseHeader handler was only setting connection close if the service was
in graceful shutdown. If not then it was assuming the service to be healthy and
sending in http request downstream. This can potentially lead to leaking of incoming http
requests to the backed end service for states other than graceful shutdown.

Adding "ServiceState" support which will add more context if and when the service
is in graceful shutdown regardless of the request being GET or POST. This will also
include support for setting the state of the service as 503 (ServiceUnavailable).
This will allow the backend service to reject any incoming http request when ALB
still thinks the service is healthy due to the fact that refresh is not yet kicked in.

One place where this service state management can come in handy is when the backend service
memory utilization is very high (~80-90%). If your state is not set to 503 then ALB will keep
on sending requests to the backend service, which just adds to the memory pressure.
The only way to protect the service is by removing from ALB which can again take up to couple
of seconds and can potentially lead to SIGKILL. If a backend web server sets using "ServiceState"
sets the state to 503 then even if ALB has not yet removed the instance, "ServiceState" as part of
web middle layers will cut off all incoming http requests, giving enough time for the backend
web server to relieve memory pressure. It's better to protect the backend service by sending 503;
instead of OOM'ing continuously and loosing customer request.

Test: Added test cases to handle the different GET & POST scenarios for all 3 states of
the service