add support to set up watcher via update function
If a user of ClientConfigChangerSink doesn't want auth watcher, then
currently there is no way to do so. Hence, introduced a new API
SetupSinkClientChanger which will accept init ClientConfigChangerSink
with the authToken that is provided. If the users later on wants to
update auth token it can be done using API AuthUpdate.

Above changes give freedom to caller for not setting auth watcher which
in some cases might not be required at all